move code + cleanup
