+++ src/ui/src/components/shaves/shave-columns.tsx
@@ -1,25 +1,11 @@
 import type { ColumnDef } from "@tanstack/react-table";
-import { Video } from "lucide-react";
-import { getYouTubeThumbnail, timeAgo } from "@/lib/utils";
+import { getStatusVariant } from "@/lib/shave-utils";
+import { timeAgo } from "@/lib/utils";
 import { SortableHeader } from "../ui/sortable-header";
 import { Badge } from "../ui/badge";
-import type { BadgeVariant, ShaveItem } from "../../types";
+import type { ShaveItem } from "../../types";
 import { ShaveAction } from "./ShaveAction";
-
-const getStatusVariant = (status: string): BadgeVariant => {
-  switch (status) {
-    case "Completed":
-      return "success";
-    case "Cancelled":
-      return "secondary";
-    case "Processing":
-      return "secondary";
-    case "Failed":
-      return "destructive";
-    default:
-      return "default";
-  }
-};
+import { VideoThumbnail } from "./VideoThumbnail";
 
 export const shaveColumns: ColumnDef<ShaveItem>[] = [
   {
@@ -28,26 +14,7 @@ export const shaveColumns: ColumnDef<ShaveItem>[] = [
     size: 120,
     enableSorting: false,
     enableColumnFilter: false,
-    cell: ({ row }) => {
-      const shave = row.original;
-      const thumbnail = shave.videoEmbedUrl ? getYouTubeThumbnail(shave.videoEmbedUrl) : null;
-      const videoUrl = shave.videoEmbedUrl?.replace("embed/", "watch?v=") || null;
-      const Wrapper = videoUrl ? "a" : "div";
-      return (
-        <Wrapper
-          className={`w-[100px] h-[56px] rounded bg-black/40 border border-white/10 flex items-center justify-center overflow-hidden ${videoUrl ? "cursor-pointer" : "cursor-default opacity-50"}`}
-          {...(videoUrl
-            ? { href: videoUrl, target: "_blank", rel: "noopener noreferrer", title: "Open video" }
-            : {})}
-        >
-          {thumbnail ? (
-            <img src={thumbnail} alt={shave.title} className="w-full h-full object-cover" />
-          ) : (
-            <Video className="h-5 w-5 text-muted-foreground" />
-          )}
-        </Wrapper>
-      );
-    },
+    cell: ({ row }) => <VideoThumbnail shave={row.original} />,
   },
   {
     accessorKey: "title",
